+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -6,6 +6,12 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* The library is now marked as trim and AOT compatible.
+  ([#7441](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7441))
+
+* The `Resource` Schema URL is now printed alongside the resource attributes.
+  ([#7472](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7472))
+
 ## 1.16.0
 
 Released 2026-Jun-10